# openapi.yaml
openapi: 3.0.2
# https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md
servers:
  - url: 'http://localhost:8080'
info:
  version: 1.0.0
  title: Scheduler Service
  description: |
    REST API for the Scheduler Service. This API allows create and manage
    scheduled jobs.
  contact:
    name: Andriy Kornatskyy
    email: andriy.kornatskyy@live.com
  license:
    name: MIT
tags:
  - name: collections
    description: Provides operations for creating and managing collections of scheduled jobs.
  - name: jobs
    description: Provides operations for creating and managing scheduled jobs.
paths:
  /collections:
    get:
      summary: Retrieves a list of job collections
      operationId: ListCollections
      tags:
        - collections
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CollectionItem'
                required:
                  - items
    post:
      summary: Creates a job collection
      operationId: CreateCollection
      tags:
        - collections
      requestBody:
        description: a job collection to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Collection'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                type: string
                format: uuid
        '400':
          description: validation errors
        '404':
          description: not found
        '409':
          description: conflict
        '422':
          description: unprocessable entity
  /collections/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      summary: Retrieves a specified job collection
      operationId: RetrieveCollection
      tags:
        - collections
      parameters:
        - $ref: '#/components/parameters/if-none-match'
      responses:
        '200':
          description: ok
          headers:
            ETag:
              $ref: '#/components/schemas/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '304':
          description: not modified
        '400':
          description: validation errors
        '404':
          description: not found
    patch:
      summary: Updates a job collection
      operationId: UpdateCollection
      tags:
        - collections
      parameters:
        - $ref: '#/components/parameters/if-match'
      requestBody:
        description: a job collection to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Collection'
      responses:
        '204':
          description: updated
        '400':
          description: validation errors
        '404':
          description: not found
        '409':
          description: conflict
        '412':
          description: precondition failed
        '422':
          description: unprocessable entity
  /jobs:
    get:
      summary: Retrieves a list of jobs
      operationId: ListJobs
      tags:
        - jobs
      parameters:
        - $ref: '#/components/parameters/collection_id'
        - name: state
          in: query
          required: false
          description: comma delimited list of state to filter by e.g. state=enabled,disabled
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobItem'
                required:
                  - items
components:
  schemas:
    CollectionItem:
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: unique job collection name
          example: my-app
          minimum: 3
          maximum: 64
        state:
          type: string
          default: enabled
          enum:
            - enabled
            - disabled
    Collection:
      allOf:
        - $ref: '#/components/schemas/CollectionItem'
      properties:
        updated:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
    JobItem:
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          description: unique job name
          readOnly: true
          example: hourly-job
        state:
          type: string
          readOnly: true
          enum:
            - enabled
            - disabled
        schedule:
          type: string
          readOnly: true
          example: '@every 1h'
  parameters:
    id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    collection_id:
      in: query
      name: collection_id
      required: false
      schema:
        type: string
        format: uuid
    if-match:
      in: header
      name: If-Match
      required: false
      schema:
        $ref: '#/components/schemas/ETag'
    if-none-match:
      in: header
      name: If-None-Match
      required: false
      schema:
        $ref: '#/components/schemas/ETag'